Adaptive sampling (#85)
* Add: Adaptive settings class

* Bugfix: use the correct init ws for AIW

was just using the entire array, so just unravelling and taking the
first few values...

* Format: trim trailing whitespace

* Add: First draft of adaptive sampling function

* Doc: Improve documentation for AIS distribution

* Add: Method to extend the convex hull

reflects values in x and y so that the whole domain becomes interior

* Add: Franke function for test development

* Refactor: Extract structured interpolation method

* Add: Unstructured interpolation

* Update: Allow two-step unstructured interpolation

* Format: trailing whitespace

* Update: Change default interpolation method

* Add: Distribution method option, just for AIS

* Bugfix: Add method to interpolate the unstructured WS

The locations are unstructured, so we can't use the old interp method
This one is just nearest neighbour and seems sufficient

* Bugfix: remove old Distribution initialisation

* Bugfix: Make sure that init N windows evals correctly for one iter

* Update: Utilise the two step interpolation

* Update: Also return the final distribution from analysis

* Bugfix: Allow returning of dist from adapt analysis

* Bugfix: Correct the order of x and y

* Bugfix: Fix n_window calculation for AIS

* Add: Cython method to get the local spatial variance

* Update: Use cython code to calc the local spatial var

Using the SATs is incorrect.

For each pixel, we need to compare local values to the CURRENT pixel.
Using the SAT effectively takes the average of each pixels delta to
_that_ pixels local mean.

Using the cython code requires the arrays to be contiguous as this
offers quite a significant (~3x) speed up

* Minor bugfix: make sure mask is always float

* Update: Allow calc of var with even spacing

* Bugfix: ensure no values are less than 0

* Add: Basic objective function calculation

* Bugfix: struc_adapt n windows bug

* Update: Extract unstructured interpolation into own function

* Add: Inverse distance weighting vector validation

* Bugfix: ensure object is copied

The id was simply being copied which was causing knock on effects.

Unusually, ensR.u.mean is u_mean_var.mean was returning false, yet the
id's were identical

+++ PIV/piv_image.py
@@ -452,7 +452,7 @@ def load_mask(flowtype):
         mask = np.asarray(Image.open(filenames[2]).convert('L')).copy()
         mask[mask > 0] = 1
 
-    return mask
+    return mask.astype(np.float64)
 
 
 def quintic_spline_image_filter(IA):
